+++ m1fp/publickey.go
@@ -12,8 +12,8 @@ type PublicKey struct {
 	XInt *big.Int // X lifted to common domain: ⌊X · D⌋
 	HInt *big.Int // H lifted to common domain: ⌊H · D⌋
 	D    *big.Int // Common denominator: 2^P · 5^n
-	Prec uint     // P: binary precision in bits
-	N    uint     // n: decimal precision (digits)
+	Prec uint16   // P: binary precision in bits
+	N    uint16   // n: decimal precision (digits)
 }
 
 // MarshalBinary encodes the public key into a compact binary representation.
@@ -32,8 +32,8 @@ func (pk *PublicKey) MarshalBinary() ([]byte, error) {
 	// Format: [prec:2][n:2][xLen:4][hLen:4][xBytes][hBytes]
 	buf := make([]byte, 12+len(xBytes)+len(hBytes))
 
-	binary.BigEndian.PutUint16(buf[0:2], uint16(pk.Prec))
-	binary.BigEndian.PutUint16(buf[2:4], uint16(pk.N))
+	binary.BigEndian.PutUint16(buf[0:2], pk.Prec)
+	binary.BigEndian.PutUint16(buf[2:4], pk.N)
 	binary.BigEndian.PutUint32(buf[4:8], uint32(len(xBytes)))
 	binary.BigEndian.PutUint32(buf[8:12], uint32(len(hBytes)))
 
@@ -62,8 +62,8 @@ func (pk *PublicKey) UnmarshalBinary(data []byte) error {
 	xBytes := data[12 : 12+xLen]
 	hBytes := data[12+xLen : 12+xLen+hLen]
 
-	pk.Prec = uint(prec)
-	pk.N = uint(n)
+	pk.Prec = uint16(prec)
+	pk.N = uint16(n)
 	pk.XInt = new(big.Int).SetBytes(xBytes)
 	pk.HInt = new(big.Int).SetBytes(hBytes)
 	pk.D = computeCommonDenominator(pk.Prec, pk.N)
@@ -73,8 +73,8 @@ func (pk *PublicKey) UnmarshalBinary(data []byte) error {
 
 // computeCommonDenominator calculates D = 2^P · 5^n for the unified arithmetic domain.
 // This denominator ensures exact conversions between binary and decimal representations.
-func computeCommonDenominator(p, n uint) *big.Int {
-	twoPowP := new(big.Int).Lsh(big.NewInt(1), p)
+func computeCommonDenominator(p, n uint16) *big.Int {
+	twoPowP := new(big.Int).Lsh(big.NewInt(1), uint(p))
 	fivePowN := new(big.Int).Exp(big.NewInt(5), big.NewInt(int64(n)), nil)
 	return new(big.Int).Mul(twoPowP, fivePowN)
 }
@@ -82,13 +82,13 @@ func computeCommonDenominator(p, n uint) *big.Int {
 // liftToCommonDomain converts a fractional value f ∈ [0,1) to the common domain
 // by computing ⌊f · D⌋ where D is the common denominator.
 // This operation preserves precision by avoiding intermediate rounding.
-func liftToCommonDomain(f *big.Float, d *big.Int, prec uint) (*big.Int, error) {
-	if f.Sign() < 0 || f.Cmp(big.NewFloat(1).SetPrec(f.Prec())) >= 0 {
+func liftToCommonDomain(f *big.Float, d *big.Int, prec uint16) (*big.Int, error) {
+	if f.Sign() < 0 || f.Cmp(big.NewFloat(1).SetPrec(uint(f.Prec()))) >= 0 {
 		return nil, errors.New("value not in [0,1)")
 	}
 
-	dFloat := new(big.Float).SetPrec(prec).SetInt(d)
-	scaled := new(big.Float).SetPrec(prec).Mul(f, dFloat)
+	dFloat := new(big.Float).SetPrec(uint(prec)).SetInt(d)
+	scaled := new(big.Float).SetPrec(uint(prec)).Mul(f, dFloat)
 	result, _ := scaled.Int(nil)
 	return result, nil
 }
